# Make ASCII case conversions more than 4× faster