--- a/src/libcore/benches/lib.rs
+++ b/src/libcore/benches/lib.rs
@@ -5,7 +5,7 @@ extern crate core;
 extern crate test;
 
 mod any;
-mod ascii_case;
+mod ascii;
 mod char;
 mod hash;
 mod iter;