@ollie27 2019-03-18
Might this be slower for platforms without SIMD which can't take advantage of auto-vectorization or does that not matter?