@SimonSapin 2019-03-18
I guess it depends on whether LLVM can auto-vectorize based on "classic" `u32` operations. But either way it’s likely still faster than the current lookup table.

I also just realize that when doing one byte at a time, instead of convoluted add-then-mask to emulate comparison, we can use actual comparison to obtain a `bool`, then cast to `u8` to obtain a 1 or 0, then multiply that by a mask:

```rust
byte &= !(0x20 * (b'a' <= byte && byte <= b'z') as u8)
```

This even turns out to be slightly faster! I’ll update the PR.